Saving a file will no longer set all rotations to -45 degrees
I made it so that not only will saving a bol no longer set ALL rotations to -45 degrees but the user can also edit the rotations via xRotation and yRotation. This allows a user to edit any of the default 16 courses without messing up the starting point rotation or the rotation of any course object, such as the cannon on Donkey Kong Mountain. The formula used to calculate the rotation is arctan(yRotation / xRotation) - 90. Making the user do the math to determine which numbers to use for xRotation and yRotation is not very user friendly but it works.